cbam_project/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
google-genai>=0.2.0
